Adjusted routes for contentful

+++ backend/tracks/routes.py
@@ -58,6 +58,9 @@ class TrackDelete(Resource):
         contentful_data = request.get_json()
         track = Track.query.filter_by(contentful_id=contentful_data["entityId"]).first()
 
+        if not track:
+            return {"message": "Track does not exist"}, 404
+
         db.session.delete(track)
         db.session.commit()
